requirements.txt
Django==3.2.6
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-gis==0.17
psycopg2==2.7.7
psycopg2-binary==2.9.1
